GEODE-3719: Migrate code out of depricated packages
