+++ geode-core/src/main/java/org/apache/geode/admin/AdminDistributedSystem.java
@@ -17,6 +17,7 @@ package org.apache.geode.admin;
 import org.apache.geode.cache.DataPolicy;
 import org.apache.geode.cache.persistence.PersistentID;
 import org.apache.geode.distributed.DistributedMember;
+import org.apache.geode.management.BackupStatus;
 
 import java.io.File;
 import java.net.InetAddress;
