+++ geode-core/src/main/java/org/apache/geode/admin/BackupStatus.java
@@ -14,32 +14,14 @@
  */
 package org.apache.geode.admin;
 
-import java.util.Map;
-import java.util.Set;
-
-import org.apache.geode.cache.persistence.PersistentID;
-import org.apache.geode.distributed.DistributedMember;
-import org.apache.geode.distributed.internal.DM;
-import org.apache.geode.internal.cache.BackupUtil;
-
 /**
  * The status of a backup operation, returned by
- * {@link BackupUtil#backupAllMembers(DM, java.io.File,java.io.File)}.
+ * {@link AdminDistributedSystem#backupAllMembers(java.io.File,java.io.File)}.
  * 
  * @since GemFire 6.5
+ * @deprecated as of 7.0 use the <code><a href=
+ *             "{@docRoot}/org/apache/geode/management/package-summary.html">management</a></code>
+ *             package instead
  */
-public interface BackupStatus {
-
-  /**
-   * Returns a map of disk stores that were successfully backed up. The key is an online distributed
-   * member. The value is the set of disk stores on that distributed member.
-   */
-  Map<DistributedMember, Set<PersistentID>> getBackedUpDiskStores();
-
-  /**
-   * Returns the set of disk stores that were known to be offline at the time of the backup. These
-   * members were not backed up. If this set is not empty the backup may not contain a complete
-   * snapshot of any partitioned regions in the distributed system.
-   */
-  Set<PersistentID> getOfflineDiskStores();
+public interface BackupStatus extends org.apache.geode.management.BackupStatus {
 }
