+++ geode-core/src/main/java/org/apache/geode/management/DiskBackupStatus.java
@@ -14,91 +14,25 @@
  */
 package org.apache.geode.management;
 
-import java.util.HashMap;
 import java.util.Map;
-import java.util.Set;
-import java.util.stream.Collectors;
 
 import org.apache.geode.cache.persistence.PersistentID;
-import org.apache.geode.distributed.DistributedMember;
 
 /**
  * Composite data type used to distribute the status of disk backup operations.
  * 
  * @since GemFire 7.0
  */
-public class DiskBackupStatus {
-
-  /**
-   * Map of DistributedMember, Set<PersistentID>
-   */
-  private Map<String, String[]> backedUpDiskStores;
-
-  /**
-   * List of offline disk stores
-   */
-  private String[] offlineDiskStores;
-
+public interface DiskBackupStatus {
   /**
    * Returns a map of member names/IDs and the {@link PersistentID} of the disk stores that were
    * backed up.
    */
-  public Map<String, String[]> getBackedUpDiskStores() {
-    return backedUpDiskStores;
-  }
+  Map<String, String[]> getBackedUpDiskStores();
 
   /**
    * Returns a list of directories for the disk stores that were off-line at the time the backup
    * occurred.
    */
-  public String[] getOfflineDiskStores() {
-    return offlineDiskStores;
-  }
-
-  /**
-   * Sets the map of member names/IDs and the {@link PersistentID} of the disk stores that were
-   * backed up.
-   */
-  @SuppressWarnings("unused")
-  public void setBackedUpDiskStores(Map<String, String[]> backedUpDiskStores) {
-    this.backedUpDiskStores = backedUpDiskStores;
-  }
-
-  /**
-   * Sets the list of directories for the disk stores that were off-line at the time the backup
-   * occurred.
-   */
-  @SuppressWarnings("unused")
-  public void setOfflineDiskStores(String[] offLineDiskStores) {
-    this.offlineDiskStores = offLineDiskStores;
-  }
-
-  /**
-   * Sets the map of member names/IDs and the {@link PersistentID} of the disk stores that were
-   * backed up.
-   */
-  public void generateBackedUpDiskStores(
-      Map<DistributedMember, Set<PersistentID>> backedUpDiskStores) {
-    Map<String, String[]> diskStores = new HashMap<>();
-    backedUpDiskStores.entrySet().forEach(entry -> {
-      DistributedMember member = entry.getKey();
-      Set<PersistentID> ids = entry.getValue();
-      String[] setOfDiskStr = new String[ids.size()];
-      entry.getValue().stream().map(id -> id.getDirectory()).collect(Collectors.toList())
-          .toArray(setOfDiskStr);
-      diskStores.put(member.getId(), setOfDiskStr);
-    });
-    setBackedUpDiskStores(diskStores);
-  }
-
-  /**
-   * Sets the list of directories for the disk stores that were off-line at the time the backup
-   * occurred.
-   */
-  public void generateOfflineDiskStores(Set<PersistentID> offLineDiskStores) {
-    String[] diskStores = new String[offLineDiskStores.size()];
-    offLineDiskStores.stream().map(id -> id.getDirectory()).collect(Collectors.toList())
-        .toArray(diskStores);
-    setOfflineDiskStores(diskStores);
-  }
+  String[] getOfflineDiskStores();
 }
